// src/app/layout/common/channel/all-channels.types.ts
export interface Channel
{
    id: number;
    channel_id: number;
    profile_pic: string;
    blur_image: string;
    title: string;
    // username: string;
    // channel_name: string;
   
    followers: string;
    viewCount: string;
    videoCount: string | null;
    subscriberCount: string;
    EngagementRate: string;
    EstViews: string;
    cpm: string;
    price_view_ratio: string;
    currency: string;
    promotion_price: number;
    is_verified: 0 | 1;
    is_default: number;
    // count: number;
    is_favourite: boolean;
    bmiProfile: string;
    youTubeProfile: string;
    channel_link: string;
    is_revealed: boolean;
    is_invite: boolean;
    channel: string;
    channel_lang?: string;
    tags?:string;
    cat_percentile_1?:string;
    cat_percentile_5?:string;
    cat_percentile_10?:string;
}